update warranty management modal

+++ src/app/components/admin/pages/warranty-management/warranty-managment-modal/warranty-response-history/warranty-response-history.component.ts
@@ -51,8 +51,11 @@ import { IUser } from '../../../../../../models/user.interface';
 export class WarrantyResponseHistoryComponent implements OnInit {
   @Input() claimNo: string = '';
   @Input() claimId: number = 0;
+  @Input() isNoSave: boolean = false;
 
   angularGrid!: AngularGridInstance;
+  deletedIds: number[] = [];
+  tempIdCounter: number = -1;
   gridId = `grid-warranty-response-history-${crypto.randomUUID()}`;
   columnDefinitions: Column[] = [];
   gridOptions: GridOption = {};
@@ -87,7 +90,7 @@ export class WarrantyResponseHistoryComponent implements OnInit {
   constructor(
     private modal: NzModalService,
     private notification: NzNotificationService,
-    private responseHistoryService: ResponseHistoryService,
+    public responseHistoryService: ResponseHistoryService,
     private customerService: CustomerService,
     private userService: UserManagementService,
   ) {}
@@ -288,6 +291,23 @@ export class WarrantyResponseHistoryComponent implements OnInit {
       WarrantyClaimId: this.claimId,
     };
 
+    if (this.isNoSave) {
+      if (this.isEditMode) {
+        const index = this.dataset.findIndex((x) => x.Id === this.formData.Id);
+        if (index !== -1) {
+          this.dataset[index] = { ...payload };
+          this.dataset = [...this.dataset];
+        }
+      } else {
+        const newId = this.tempIdCounter--;
+        const newItem = { ...payload, Id: newId };
+        this.dataset = [newItem, ...this.dataset];
+      }
+      this.isModalVisible = false;
+      this.resetForm();
+      return;
+    }
+
     if (this.isEditMode) {
       this.responseHistoryService.update(this.formData.Id!, payload).subscribe({
         next: () => {
@@ -352,6 +372,15 @@ export class WarrantyResponseHistoryComponent implements OnInit {
 
     const selectedItem =
       this.angularGrid.gridService.getSelectedRowsDataItem()[0];
+
+    if (this.isNoSave) {
+      if (selectedItem.Id > 0) {
+        this.deletedIds.push(selectedItem.Id);
+      }
+      this.dataset = this.dataset.filter((x) => x.Id !== selectedItem.Id);
+      return;
+    }
+
     this.responseHistoryService.delete(selectedItem.Id).subscribe({
       next: () => {
         this.notification.success(NOTIFICATION_TITLE.success, 'Xóa thành công');
@@ -373,5 +402,17 @@ export class WarrantyResponseHistoryComponent implements OnInit {
       toDate: null,
     };
     this.loadData();
+  }
+
+  getAddedItems() {
+    return this.dataset.filter((x) => x.Id < 0);
+  }
+
+  getUpdatedItems() {
+    return this.dataset.filter((x) => x.Id > 0);
+  }
+
+  getDeletedIds() {
+    return this.deletedIds;
   }
 }
